+++ packages/extension/src/view/devtools/components/cookies/cookiesListing/useAllowedList/isCookieDomainInAllowList.ts
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * External dependencies.
+ */
+import type { CookieTableData } from '@ps-analysis-tool/common';
+
+const isCookieDomainInAllowList = (
+  cookie: CookieTableData,
+  domainsInAllowList: Set<string>
+) => {
+  const domain = cookie.parsedCookie?.domain || '';
+  let _domain = '';
+
+  if (domain) {
+    _domain = domain.startsWith('.') ? domain : `.${domain}`;
+  }
+
+  let isDomainInAllowList = domainsInAllowList.has(_domain);
+
+  if (!isDomainInAllowList) {
+    isDomainInAllowList = [...domainsInAllowList].some((storedDomain) => {
+      // For example xyz.bbc.com and .bbc.com
+      return _domain.endsWith(storedDomain) && _domain !== storedDomain;
+    });
+  }
+
+  return isDomainInAllowList;
+};
+
+export default isCookieDomainInAllowList;
